Move `FormattedJson` to devtools_app_shared
